Drop Yast::Routing from routing client.
